backend/pyproject.toml
[project]
name = "pdf-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "inngest>=0.5.15",
    "langchain-community>=0.4.1",
    "llama-index-core>=0.14.13",
    "llama-index-readers-file>=0.5.6",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.16.2",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
]
